fix: absolute paths should not use the guest cwd
This is a follow-up to #30.

Removing the `vim.loop.fs_realpath` check helps with relative paths that
exist in the guest, but it breaks absolute paths.

For example, in a neovim terminal

```shell
cd /tmp/
vim /tmp/a/file.txt
```

Would concatenate the path to `/tmp//tmp/a/file.

I'm guessing checking for the real path was supposed to handle that
scenario, it just conflicted with relative paths that matched the host
cwd.

This commit checks each file to see if it's absolute and avoid
prepending the guest cwd.